osd: Add ECTypes and add ec_align_t to it
Move out the common types for new and old EC into ECTypes. These are mainly for use in interfaces, with ECUtils reerved for utilities.

More will be added later.
