+++ src/osd/ReplicatedBackend.h
@@ -148,7 +148,7 @@ public:
 
   void objects_read_async(
     const hobject_t &hoid,
-    const std::list<std::pair<ECCommonL::ec_align_t,
+    const std::list<std::pair<ec_align_t,
 	       std::pair<ceph::buffer::list*, Context*> > > &to_read,
                Context *on_complete,
                bool fast_read = false) override;
